+++ src/bc/globals.d.ts
@@ -2,7 +2,14 @@ declare global {
   interface BCPlayer {
     MemberNumber: number;
     Name: string;
+    Nickname?: string;
     FriendNames: Map<number, string>;
+  }
+
+  interface BCCharacter {
+    MemberNumber: number;
+    Name: string;
+    Nickname?: string;
   }
 
   interface BCChatRoomData {
@@ -21,9 +28,22 @@ declare global {
     Private?: boolean;
   }
 
+  interface BCFriendListBeepLogMessage {
+    MemberNumber: number;
+    MemberName: string;
+    ChatRoomName?: string;
+    ChatRoomSpace?: string;
+    Private?: boolean;
+    Sent: boolean;
+    Time: Date | string | number;
+    Message?: string;
+  }
+
   var Player: BCPlayer;
   var CurrentScreen: string;
   var ChatRoomData: BCChatRoomData | null;
+  var ChatRoomCharacter: BCCharacter[];
+  var FriendListBeepLog: BCFriendListBeepLogMessage[];
 
   function ServerAccountBeep(data: BCServerAccountBeepResponse): void;
   function ServerSendBeepMessage(
@@ -32,6 +52,7 @@ declare global {
     options?: { includeRoom?: boolean },
   ): void;
   function ServerIsLoggedIn(): boolean;
+  function CharacterNickname(character: BCCharacter): string;
 }
 
 export {};